apps/api/pyproject.toml
[project]
name = "memllm-api"
version = "0.1.0"
description = "FastAPI API for the Letta-native memLLM runtime"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "memllm-domain",
    "memllm-letta-integration",
    "pydantic-settings>=2.8.1",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.0",
]

[project.scripts]
memllm-api = "memllm_api.cli:main"

[tool.uv.sources]
memllm-domain = { workspace = true }
memllm-letta-integration = { workspace = true }

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"
